third_party/blink/web_tests/fast/js/function-length-expected.txt

This tests the length property of functions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.confirm.length is 0
PASS window.open.length is 0
PASS window.setTimeout.length is 1
PASS window.clearTimeout.length is 0
PASS window.addEventListener.length is 2
PASS window.removeEventListener.length is 2
PASS window.postMessage.length is 1
PASS window.dispatchEvent.length is 1
PASS window.openDatabase.length is 4
PASS window.history.pushState.length is 2
PASS window.history.length is 1
PASS window.URL.createObjectURL.length is 1
PASS window.Storage.prototype.key.length is 1
PASS document.importNode.length is 1
PASS document.write.length is 0
PASS CanvasRenderingContext2D.prototype.drawImage.length is 3
PASS Array.prototype.concat.length is 1
PASS Array.prototype.join.length is 1
PASS Array.prototype.push.length is 1
PASS Array.prototype.slice.length is 2
PASS Date.UTC.length is 7
PASS Date.prototype.setSeconds.length is 2
PASS Date.prototype.setMinutes.length is 3
PASS Date.prototype.setHours.length is 4
PASS Date.prototype.setMonth.length is 2
PASS Date.prototype.setFullYear.length is 3
PASS Number.prototype.toExponential.length is 1
PASS Number.prototype.toFixed.length is 1
PASS Number.prototype.toPrecision.length is 1
PASS Math.max.length is 2
PASS Math.min.length is 2
PASS Function.prototype.apply.length is 2
PASS Function.prototype.call.length is 1
PASS String.prototype.split.length is 2
PASS String.prototype.substring.length is 2
PASS successfullyParsed is true

TEST COMPLETE
